small update

+++ src/main/java/org/onedatashare/server/model/request/UserTransferOptions.java
@@ -36,5 +36,5 @@ public class UserTransferOptions {
     private Integer concurrencyThreadCount;
     private Integer parallelThreadCount;
     private Integer pipeSize;
-    private Integer chunkSize;
+    private Long chunkSize;
 }
